Continue consolidate parts of the project
