+++ src/libraries/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_subdirectory( tcp_daq )
+
+
