+++ src/plugins/CMakeLists.txt
@@ -1,5 +1,5 @@
 
-add_subdirectory( EVIOParser )
+add_subdirectory(evio)
 add_subdirectory(example_evio_analysis)
